+++ src/main/java/de/slikey/effectlib/util/ParticleDisplay.java
@@ -27,7 +27,6 @@ public abstract class ParticleDisplay {
     protected void spawnParticle(Particle particle, ParticleOptions options, Location center, double range, List<Player> targetPlayers) {
         try {
             if (targetPlayers == null) {
-                // Note: this code is mirrored in EffectManager
                 double squared = range * range;
                 for (final Player player : Bukkit.getOnlinePlayers()) {
                     if (!manager.isVisiblePlayer(player, center, squared)) continue;